LayoutTests/imported/w3c/web-platform-tests/html/user-activation/consumption-sameorigin.tentative-expected.txt

CONSOLE MESSAGE: TypeError: undefined is not an object (evaluating 'navigator.userActivation.isActive')
CONSOLE MESSAGE: TypeError: undefined is not an object (evaluating 'navigator.userActivation.isActive')
CONSOLE MESSAGE: TypeError: undefined is not an object (evaluating 'navigator.userActivation.isActive')
User activation consumption across same-origin frame boundary

Tests that user activation consumption resets the transient states in all same-origin frames.

Click anywhere on the green area (child frame).
  

Harness Error (TIMEOUT), message = null

NOTRUN Consumption test 
